patch by don rollo (modified patch of java.util.Properties, omitted build.xml stuff)
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@3900 06c8f799-c779-4783-8af7-179afedb67c8

+++ classes/lejos/robotics/localization/MCLParticleSet.java
@@ -303,7 +303,7 @@ public class MCLParticleSet {
    if(rr.incomplete()) return false;
     maxWeight = 0f;
     totalWeight = 0f;
-    float minWeight = (float) .001f/numParticles;
+    float minWeight = .001f /numParticles;
     for (int i = 0; i < numParticles; i++) {
       particles[i].calculateWeight(rr, map, twoSigmaSquared);
       float weight = particles[i].getWeight();
